+++ app/Controllers/BaseController.php
@@ -52,7 +52,7 @@ class BaseController extends Controller
 		//--------------------------------------------------------------------
 		// Preload any models, libraries, etc, here.
 		//--------------------------------------------------------------------
-		// E.g.: $this->session = \Config\Services::session();
+		// E.g.: $this->session = \Config\Services::session();		
 		$this->data = [
 			'loggedUser' => user(),
 		];